requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.56.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.3
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
six==1.17.0
Werkzeug==3.1.3
